+++ packages/plugin/workspace/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @dovenv/workspace
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @dovenv/core@0.0.20
+
 ## 0.0.19
 
 ### Patch Changes